add comment and refine threshold choice
