+++ test/e2e/chaos/stresschaos/cpu.go
@@ -25,51 +25,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-func TestcaseMemoryStressInjectionOnceThenRecover(
-	ns string,
-	cli client.Client,
-	peers []*corev1.Pod,
-	ports []uint16,
-	c http.Client,
-) {
-	ctx := context.Background()
-	By("create memory stress chaos CRD objects")
-	memoryStressChaos := makeMemoryStressChaos(ns, "memory-stress", ns, "stress-peer-0", "50M", 1)
-	err := cli.Create(ctx, memoryStressChaos.DeepCopy())
-	framework.ExpectNoError(err, "create stresschaos error")
-
-	By("waiting for assertion some pods are experiencing memory stress ")
-	err = wait.Poll(time.Second, 15*time.Second, func() (done bool, err error) {
-		conditions, err := probeStressCondition(c, peers, ports)
-		if err != nil {
-			return false, err
-		}
-		if conditions[0].MemoryUsage > 50*1024*1024 && conditions[1].MemoryUsage < 50*1024*1024 {
-			return true, nil
-		}
-		framework.Logf("get Memory: [%d, %d]", conditions[0].MemoryUsage, conditions[1].MemoryUsage)
-		return false, nil
-	})
-	framework.ExpectNoError(err, "memory stress failed")
-	By("delete pod failure chaos CRD objects")
-
-	err = cli.Delete(ctx, memoryStressChaos.DeepCopy())
-	framework.ExpectNoError(err, "delete stresschaos error")
-	By("waiting for assertion recovering")
-	err = wait.Poll(time.Second, 15*time.Second, func() (done bool, err error) {
-		conditions, err := probeStressCondition(c, peers, ports)
-		if err != nil {
-			return false, err
-		}
-		if conditions[0].MemoryUsage < 50*1024*1024 && conditions[1].MemoryUsage < 50*1024*1024 {
-			return true, nil
-		}
-		framework.Logf("get Memory: [%d, %d]", conditions[0].MemoryUsage, conditions[1].MemoryUsage)
-		return false, nil
-	})
-	framework.ExpectNoError(err, "fail to recover from memory stress")
-}
-
 func TestcaseCPUStressInjectionOnceThenRecover(
 	ns string,
 	cli client.Client,
@@ -97,6 +52,8 @@ func TestcaseCPUStressInjectionOnceThenRecover(
 		lastCPUTime[0] = conditions[0].CpuTime
 		lastCPUTime[1] = conditions[1].CpuTime
 		framework.Logf("get CPU: [%d, %d]", diff[0], diff[1])
+		// diff means the increasing CPU time (in nanosecond)
+		// just pick two threshold, 5e8 is a little shorter than one second
 		if diff[0] > 5e8 && diff[1] < 5e6 {
 			return true, nil
 		}
@@ -121,6 +78,8 @@ func TestcaseCPUStressInjectionOnceThenRecover(
 		lastCPUTime[0] = conditions[0].CpuTime
 		lastCPUTime[1] = conditions[1].CpuTime
 		framework.Logf("get CPU: [%d, %d]", diff[0], diff[1])
+		// diff means the increasing CPU time (in nanosecond)
+		// just pick two threshold, they are both much shorter than 1 second
 		if diff[0] < 1e7 && diff[1] < 5e6 {
 			return true, nil
 		}
